+++ packages/eas-cli/src/submit/ios/__tests__/AscApiKeySource-test.ts
@@ -58,7 +58,6 @@ async function getIosSubmissionContextAsync(): Promise<SubmissionContext<Platfor
     },
     nonInteractive: true,
     isVerboseFastlaneEnabled: false,
-    groups: [],
     actor: mockJester,
     graphqlClient,
     analytics,